+++ r-nat-gateway.tf
@@ -7,6 +7,8 @@ resource "azurerm_public_ip" "pip" {
   zones               = var.public_ip_zones
   sku                 = "Standard"
 
+  domain_name_label = var.public_ip_domain_name_label
+
   tags = merge(
     local.default_tags,
     var.extra_tags