+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Added
+  * AZ-XXX : Add public ip Domain dns label option
+
 # v4.1.0 - 2022-04-15
 
 Added